dataset_build: pre-build prep — fivek/ppr10k integration, tag+SAM3 caches, stage0, fixes
Autonomous pre-build run (2026-06-03). Source pipeline is now cache-backed and
the source set expanded to ~1M with two new real-data streams.

Data integration:
- S8 fivek gold global stream (monetGPT fivek_mmart_like/train_global): real
  before->expert-after JPG (after_source=REAL_JPG) + instructions; 16,372 pairs;
  test_global (5k) held out for eval.
- S4 revived (monetGPT ppr10k): real human masks + per-source target XMP rendered
  by the teacher (PARAM), 3 experts x 8875 = 26,625.
- Budgets rebalanced to exactly 1,000,000; source_index = 124,459.

Precompute caches (read when use_cache flags are on, now true):
- tag_precompute.py / tag_cache.py / aesthetic.py: per-source scene/style/concepts
  + VLM aesthetic + LAION CLIP+MLP aesthetic; tag_cache = 106,649 entries
  (Qwen3.5-4B dual-card vLLM).
- SAM3 cache (81,327) precomputed per-image from each image's tag concepts so
  CachedMasker hits the build's dynamic queries.

Fixes:
- masking.py: cast scores bf16->float32 before .numpy() (SAM3 had never run; latent
  crash since use_cache was always false).
- sam3_precompute.py: derive per-image concepts from tag_cache (aligns with build,
  far fewer forwards) + --from-tags/--max-concepts.
- config: vllm -> Qwen3.5-4B; tag_cache + aesthetic blocks; sam3/tag use_cache=true;
  er_recon_psnr_min 25.0 -> 22.0 (stage0: floor 22 keeps ~80% of S1/S7 vs 60% at 25).

stage0 S1/S7 z* drop-rate (60-sample, post forward-fix): 40% drop @ floor 25,
~20% @ floor 22; PSNR mean 27.0.

+++ dataset_build/masking.py
@@ -455,7 +455,7 @@ class Sam3Masker(ConceptMasker):
 
         # Filter instances by score.
         if scores is not None:
-            sc = scores.detach().cpu().numpy() if hasattr(scores, "detach") else np.asarray(scores)
+            sc = scores.detach().to(torch.float32).cpu().numpy() if hasattr(scores, "detach") else np.asarray(scores)
             keep = sc >= float(min_score)
             if keep.any():
                 m = m[keep]
